+++ README.md
@@ -23,7 +23,7 @@ $ python setup.py install
 Now you may install `pylexibank` via pip or in development mode following the instructions
 in [CONTRIBUTING.md](CONTRIBUTING.md).
 
-Some functionality in `pylexibank` (in particular the `install` sub-command), require access
+Some functionality in `pylexibank` (in particular the `makecldf` sub-command), require access
 to [Glottolog](http://glottolog.org) or [Concepticon](http://concepticon.pylexibank.org) data.
 Since the data of both these applications is curated in git repositories as well, the easiest
 way to achieve this is by cloning [pylexibank/glottolog](https://github.com/pylexibank/glottolog) and
@@ -35,7 +35,7 @@ way to achieve this is by cloning [pylexibank/glottolog](https://github.com/pyle
 
 `pylexibank` can be used in two ways:
 - The command line interface provides mainly access to the functionality for the `lexibank`
-  [curation workflow](workflow.md).
+  curation workflow.
 - The `pylexibank` package can also be used like any other python package in your own
   python code to access lexibank data in a programmatic (and consistent) way.
