+++ tests/test_transcription.py
@@ -1,0 +1,18 @@
+# coding: utf8
+from __future__ import unicode_literals, print_function, division
+
+import pytest
+
+from pylexibank.transcription import analyze, Analysis
+
+
+def test_analyze():
+    with pytest.raises(ValueError):
+        analyze([], Analysis())
+
+    with pytest.raises(ValueError):
+        analyze(['\n'], Analysis())
+
+    segments, la, clpa, analysis = analyze(['a', '^', 'b'], Analysis())
+    assert segments == ['a', '^', 'b']
+    assert analysis.general_errors == 1
